refactor : molecules
