+++ src/components/atoms/SquareBtn/index.js
@@ -1,0 +1,1 @@
+export { SquareBtn } from "./SquareBtn";
