+++ src/components/molecules/MyFooter/MyFooter.jsx
@@ -1,9 +1,15 @@
 import React, { useContext } from "react";
 import { useNavigate } from "react-router-dom";
-import * as S from "./style";
+import * as S from "./MyFooter.styles";
 
 import AtuhContext from "store/auth-context";
 
+/**
+ * My page의 하단 부분입니다.
+ * Logout 버튼을 누르면 로그아웃이 되고, 메인 페이지로 이동합니다.
+ *
+ * @returns
+ */
 export const MyFooter = () => {
   const authCtx = useContext(AtuhContext);
   const navigate = useNavigate();
