+++ src/components/molecules/MyFooter/index.js
@@ -1,0 +1,1 @@
+export { MyFooter } from "./MyFooter";
