--- src/components/molecules/TodoList/TodoTemplate.jsx
@@ -1,14 +1,0 @@
-import React from "react";
-import styled from "styled-components";
-
-const TodoTemplate = ({ children }) => {
-  return <TodoTemplateBlock>{children}</TodoTemplateBlock>;
-};
-
-export default TodoTemplate;
-
-const TodoTemplateBlock = styled.div`
-  width: 100%;
-  padding: 24px;
-  box-sizing: border-box;
-`;
