+++ src/components/molecules/Week/index.js
@@ -1,0 +1,1 @@
+export { Week } from "./Week";
