+++ src/components/organisms/TodoList/TodoList.jsx
@@ -1,8 +1,5 @@
-import React, { useState, useRef } from "react";
-import ToDoInsert from "./ToDoInsert";
-import TodoListBlock from "./TodoListBlock";
-import TodoTemplate from "./TodoTemplate";
-
+import { ToDoInsert, TodoListBlock } from "components";
+import * as S from "./TodoList.styles";
 /**
  * TodoList
  *
@@ -16,11 +13,16 @@ import TodoTemplate from "./TodoTemplate";
  * @returns {JSX.Element} TodoInsert Todo 추가 컴포넌트
  * @returns {JSX.Element} TodoListBlock todo 리스트를 보여주는 컴포넌트
  */
-export const TodoList = ({ todos, onInsert, onRemove, onEdit }) => {
+
+const TodoTemplate = ({ children }) => {
+  return <S.TodoTemplateBlock>{children}</S.TodoTemplateBlock>;
+};
+
+export function TodoList({ todos, onInsert, onRemove, onEdit }) {
   return (
     <TodoTemplate>
       <ToDoInsert onInsert={onInsert} />
       <TodoListBlock todos={todos} onRemove={onRemove} onEdit={onEdit} />
     </TodoTemplate>
   );
-};
+}
