+++ src/components/organisms/TodoList/TodoList.styles.jsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const TodoTemplateBlock = styled.div`
+  width: 100%;
+  padding: 24px;
+  box-sizing: border-box;
+`;
