+++ src/constants/index.js
@@ -1,4 +1,4 @@
 export const POMODORO_TIME = 25;
 // export const DLAY_TIME = 600; // Test delay time
-export const DLAY_TIME = 60000; // deployment delay time
-export const ELLIPSISLENGTH = 20;
+export const DLAY_TIME = 600; // deployment delay time
+export const ELLIPSISLENGTH = 15;
